ml-api/requirement.txt
streamlit==1.25.0
tensorflow==2.12.0
pillow==10.0.0
numpy==1.24.3